test(audience): convert tests to tsx and add full integration tests
